+++ pages/index.tsx
@@ -157,7 +157,7 @@ const Home = () => {
   };
 
   return(
-    <div className="flex flex-col items-center pb-3 h-[calc(100vh-10rem)] bg-gray-300">
+    <div className="flex flex-col items-center h-[calc(100vh-10rem)] bg-gray-300">
     {/* 앱전용 화면크기 <div className="flex flex-col items-center pb-3 h-[calc(100vh-6.5rem)] bg-gray-300"> */}
 
       <h1 className="font-medium text-4xl text-center mt-8">간단하게 만드는</h1>
@@ -200,13 +200,14 @@ const Home = () => {
         </div>
       </div>
 
-      <div className="w-[24rem] h-48 mt-[2rem]">
+      <div className="w-[24rem] h-48 mt-[2rem] bg-gray-300">
         {user ? 
           <div className="flex flex-col w-[24rem] h-48 items-center">
             {/* 이제 하단에 Stepper형식의 서비스 생성코드 추가필요. */}
             
-            <div className="mt-[5rem]">
-            <CreateServiceStepper />
+            <div className="">
+              <h2 className="text-2xl font-bold text-center mb-1">나만의 명함 만들기</h2>
+              <CreateServiceStepper />
             </div>
