+++ public/d5pfr8/data.json
@@ -1,0 +1,64 @@
+{
+    "items": [
+        {
+            "icon": "github",
+            "id": "1",
+            "link": "https://github.com/Kim-Geonwoo",
+            "title": "링크 바로가기",
+            "type": "link"
+        },
+        {
+            "description": "kgw@geonwoo.dev",
+            "icon": "envelope",
+            "id": "2",
+            "link": "kgw@geonwoo.dev",
+            "title": "이메일",
+            "type": "email"
+        },
+        {
+            "icon": "instagram",
+            "id": "3",
+            "link": "https://www.instagram.com/geonwoo_music",
+            "title": "INSTAGRAM",
+            "type": "link"
+        },
+        {
+            "icon": "apple",
+            "id": "4",
+            "link": "https://music.apple.com/us/artist/%EA%B9%80%EA%B1%B0%EB%88%84/1774774538",
+            "title": "Apple Music",
+            "type": "link"
+        },
+        {
+            "icon": "spotify",
+            "id": "5",
+            "link": "https://t.ly/odhmM",
+            "title": "Spotify",
+            "type": "link"
+        },
+        {
+            "icon": "phone",
+            "id": "6",
+            "link": "010-1234-5678",
+            "title": "전화하기",
+            "type": "tel"
+        },
+        {
+            "icon": "comment",
+            "id": "7",
+            "link": "010-1234-5678",
+            "title": "문자 메시지 보내기",
+            "type": "sms"
+        },
+        {
+            "id": "8",
+            "title": "설명글_1",
+            "content": [
+                "안녕하세요.",
+                "🎸 기타 아이콘을 가지고있는",
+                "🎵 노래듣는 김건우 입니다"
+            ],
+            "type": "description"
+        }
+    ]
+  }
