+++ styles/stepper.css
@@ -5,7 +5,6 @@
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    padding: 1rem;
   }
   
   @media (min-width: 640px) {
